Refactor Server.cpp: Add V8 engine initialization and disposal methods to Server_V8.cpp
